packages/cli: use react-dev-utils to select port and open browser + bit of refactor
