+++ frontend/packages/cli/package.json
@@ -36,6 +36,7 @@
     "handlebars": "^4.7.3",
     "html-webpack-plugin": "^3.2.0",
     "inquirer": "^7.0.4",
+    "react-dev-utils": "^10.2.0",
     "recursive-readdir": "^2.2.2",
     "replace-in-file": "^5.0.2",
     "ts-loader": "^6.2.1",
